refactor(MVVM): Add packages and viewmodel to start applying MVVM architecture
