+++ app/src/main/java/com/example/kotlin/newapp/framework/adapters/viewholders/MovieViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.kotlin.mypokedexapp
+package com.example.kotlin.newapp.framework.adapters.viewholders
 
 import android.content.Context
 import android.util.Log
@@ -8,9 +8,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.Priority
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.request.RequestOptions
-import com.example.kotlin.newapp.data.movie.BaseMovie
-import com.example.kotlin.newapp.data.network.model.MovieObject
-import com.example.kotlin.newapp.data.repository.MovieRepository
+import com.example.kotlin.newapp.data.network.model.BaseMovie
 import com.example.kotlin.newapp.databinding.ItemMovieBinding
 import com.example.kotlin.newapp.utils.Constants
 import kotlinx.coroutines.CoroutineScope
@@ -18,7 +16,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class MovieViewHolder(private val binding: ItemMovieBinding) : RecyclerView.ViewHolder(binding.root) {
-    fun bind(item: BaseMovie,context:Context){
+    fun bind(item: BaseMovie, context:Context){
         binding.TVName.text = item.title
         Log.d("poster",item.poster_path)
         getMovieInfo(item.poster_path,binding.IVPhoto,context)
